move shared GPU graph infrastructure from extensions into src/
