+++ SpeedyTransforms/src/SpeedyTransforms.jl
@@ -68,6 +68,7 @@ include("array_utils.jl")
 include("spectral_transform.jl")
 include("matrix_transform.jl")
 include("fourier.jl")
+include("fourier_gpu_graphs.jl")
 include("legendre.jl")
 include("legendre_ka.jl")
 include("spectral_gradients.jl")
